requirements.txt
python-telegram-bot==21.7
openai>=1.57.0
requests==2.31.0
python-dotenv==1.0.1
sqlalchemy>=2.0.36
alembic==1.13.1
psycopg[binary]>=3.2.2
google-api-python-client==2.152.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

